ref(onboarding): Split winforms onboarding docs (#102938)
Contributes to
https://linear.app/getsentry/issue/TET-864/introduce-folders-for-onboarding-platforms
