+++ static/app/gettingStartedDocs/dotnet/winforms/onboarding.spec.tsx
@@ -4,7 +4,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './winforms';
+import docs from './index';
 
 describe('winforms onboarding docs', () => {
   it('renders errors onboarding docs correctly', async () => {
